anadido properties para varias cosas

+++ src/view/FunctionListUI.xaml.cs
@@ -8,6 +8,7 @@ using System.Windows.Shapes;
 using IGUWPF.src.models.ViewModel;
 using IGUWPF.src.bean;
 using System.Windows.Controls;
+using IGUWPF.src.utils;
 
 namespace IGUWPF.src.view.Windows
 {
@@ -29,6 +30,8 @@ namespace IGUWPF.src.view.Windows
             FunctionComboBox.ItemsSource = InitializeFunctionComboBox();
             FunctionComboBox.SelectionChanged += EditFunction_ShowAndHideCValue;
 
+            Console.WriteLine(ViewModel.XMin);
+
             //Add handlers 
             //Handlers for button events
             SaveFileButton.Click += SaveProject;
@@ -46,12 +49,14 @@ namespace IGUWPF.src.view.Windows
         private void SaveProject(object sender, RoutedEventArgs e)
         {
             //Show dialog to choose the path to save the project
-            SaveFileDialog SaveFileForm = new SaveFileDialog();
-            SaveFileForm.Title = "Save project";
-            SaveFileForm.FileName = "Desktop"; // Default file name
-            SaveFileForm.DefaultExt = ".maclab"; // Default file extension
-            SaveFileForm.Filter = "MacLab Project (." + Constants.ProjectFileExtension + ")|*." + Constants.ProjectFileExtension;
-            SaveFileForm.AddExtension = true;
+            SaveFileDialog SaveFileForm = new SaveFileDialog()
+            {
+                Title = "Save project",
+                FileName = "Desktop",
+                DefaultExt = ".maclab",
+                Filter = "MacLab Project (." + Constants.ProjectFileExtension + ")|*." + Constants.ProjectFileExtension,
+                AddExtension = true
+            };
 
             Nullable<bool> result = SaveFileForm.ShowDialog();
             if (false == result)
@@ -61,7 +66,7 @@ namespace IGUWPF.src.view.Windows
             result = ViewModel.ExportModel(SaveFileForm.FileName);
             if (result == false)
             {
-                MessageBox.Show(Constants.IOErrorMsg, Constants.ErrorWindowTitle, MessageBoxButton.OK, MessageBoxImage.Error);
+                MessageBox.Show(LanguageProperties.IOErrorMsg, LanguageProperties.ErrorWindowTitle, MessageBoxButton.OK, MessageBoxImage.Error);
                 return;
             }
         }
@@ -69,12 +74,14 @@ namespace IGUWPF.src.view.Windows
         private void OpenProject(object sender, RoutedEventArgs e)
         {
             //Show dialog to choose the project to import
-            OpenFileDialog OpenFileForm = new OpenFileDialog();
-            OpenFileForm.FileName = "Open project";
-            OpenFileForm.FileName = "Desktop"; // Default file name
-            OpenFileForm.DefaultExt = ".maclab"; // Default file extension
-            OpenFileForm.Filter = "MacLab Project (." + Constants.ProjectFileExtension + ")|*." + Constants.ProjectFileExtension;
-            OpenFileForm.Multiselect = false;
+            OpenFileDialog OpenFileForm = new OpenFileDialog()
+            {
+                Title = "Open project",
+                FileName = "Desktop",
+                DefaultExt = ".maclab",
+                Filter = "MacLab Project (." + Constants.ProjectFileExtension + ")|*." + Constants.ProjectFileExtension,
+                Multiselect = false
+            };
 
             Nullable<bool> result = OpenFileForm.ShowDialog();
             if (false == result)
@@ -84,7 +91,7 @@ namespace IGUWPF.src.view.Windows
             result = ViewModel.ImportModel(OpenFileForm.FileName);
             if (result == false)
             {
-                MessageBox.Show(Constants.IOErrorMsg, Constants.ErrorWindowTitle, MessageBoxButton.OK, MessageBoxImage.Error);
+                MessageBox.Show(LanguageProperties.IOErrorMsg, LanguageProperties.ErrorWindowTitle, MessageBoxButton.OK, MessageBoxImage.Error);
                 return;
             }
         }
@@ -94,12 +101,14 @@ namespace IGUWPF.src.view.Windows
             RepresentationParameters RepresentationParamters = ViewModel.RepresentationParameters;
 
             //Show dialog to edit de properties
-            SettingsForm SettingsForm = new SettingsForm();
-            //Load older values
-            SettingsForm.Xmin = RepresentationParamters.XMin;
-            SettingsForm.Xmax = RepresentationParamters.XMax;
-            SettingsForm.Ymin = RepresentationParamters.YMin;
-            SettingsForm.Ymax = RepresentationParamters.YMax;
+            SettingsForm SettingsForm = new SettingsForm()
+            {
+                //Load older values
+                Xmin = RepresentationParamters.XMin,
+                Xmax = RepresentationParamters.XMax,
+                Ymin = RepresentationParamters.YMin,
+                Ymax = RepresentationParamters.YMax
+            };
 
             SettingsForm.ShowDialog();
             if (false == SettingsForm.DialogResult)
@@ -130,8 +139,12 @@ namespace IGUWPF.src.view.Windows
         private void DeleteFunction(object sender, RoutedEventArgs e)
         {
             Function Function = (Function)FunctionDataGrid.SelectedItem;
-            if(null != Function)
-                ViewModel.DeleteElement(Function);
+            if (null != Function)
+            {
+                MessageBoxResult Result = MessageBox.Show(LanguageProperties.DeleteConfirmationMsg, LanguageProperties.DeleteConfirmationWindowTitle,MessageBoxButton.OKCancel,MessageBoxImage.Warning);
+                if(Result == MessageBoxResult.OK)
+                    ViewModel.DeleteElement(Function);
+            }
         }
 
         private void EditFunction(object sender, EventArgs e)
@@ -142,10 +155,12 @@ namespace IGUWPF.src.view.Windows
                 return;
 
             //Display formulary
-            ExpressionSelectorUI Form = new ExpressionSelectorUI();
-            Form.Title = "Editar expresion";
+            ExpressionSelectorUI Form = new ExpressionSelectorUI()
+            {
+                Title = "Editar expresion",
+                Calculator = Function.Calculator
+            };
             Form.FunctionComboBox.ItemsSource = InitializeFunctionComboBox();
-            Form.Calculator = Function.Calculator;
 
             Form.ShowDialog();
             if (false == Form.DialogResult)
@@ -199,16 +214,15 @@ namespace IGUWPF.src.view.Windows
         }
 
         private Function TakeFunctionDataFromAddFunctionForm() {
-            double toTest;
 
-            if (!double.TryParse(AValueTextBox.Text, out toTest) ||
+            if (!double.TryParse(AValueTextBox.Text, out double toTest) ||
                 !double.TryParse(BValueTextBox.Text, out toTest) ||
                 !double.TryParse(CValueTextBox.Text, out toTest) ||
                 FunctionNameTextBox.Text.Length == 0 ||
                 ColorSelector.SelectedColor == null ||
                 FunctionComboBox.SelectedIndex == -1)
             {
-                MessageBox.Show(Constants.IncorrectDataMsg, Constants.ErrorWindowTitle, MessageBoxButton.OK, MessageBoxImage.Error);
+                MessageBox.Show(LanguageProperties.IncorrectDataMsg, LanguageProperties.ErrorWindowTitle, MessageBoxButton.OK, MessageBoxImage.Error);
                 return null;
             }
